refactor: layout파일 내용을 여러 component들과 hooks로 분리중
