+++ src/app/index.tsx
@@ -1,4 +1,4 @@
-import { KanbanLayout } from '@/features/kanban/KanbanLayout';
+import { KanbanLayout } from '@/features/kanban/components/KanbanLayout';
 
 export function App() {
   return (
