+++ src/features/kanban/constants/kanban.ts
@@ -1,5 +1,5 @@
 export enum E_Team {
-  all = '전체',
+  all = 'All',
   lead = 'Lead',
   pl = 'PL',
   fe = 'FE',
@@ -69,6 +69,18 @@ export const netteeMembers = {
     '이성훈',
   ],
 } as const;
+
+export const sidebarList = ['project', 'team', 'assignee', 'label', 'more'];
+export const projectList = ['All', 'Blolet', 'Kanban', 'onBoard'];
+export const dummyLabels = [
+  '보류',
+  '낮음',
+  '보통',
+  '보통',
+  '높음',
+  '높음',
+  '매우 높음',
+];
 
 export const kanbanStyleMap = {
   TODO: {
